backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
limits==5.4.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
